# src/swagger-doc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The data to create new product",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              },
              "examples": {
                "single": {
                  "value": {
                    "name": "men shoes",
                    "categoryId": 5,
                    "unit": "1 pair",
                    "description": "1 pair of leather men shoes",
                    "type": "HISTORICAL"
                  }
                },
                "multiple": {
                  "value": [
                    {
                      "name": "men shoes",
                      "categoryId": 5,
                      "unit": "1 pair",
                      "description": "1 pair of leather men shoes",
                      "type": "HISTORICAL"
                    },
                    {
                      "name": "internet",
                      "categoryId": 6,
                      "unit": "monthly subscription",
                      "description": "1 month internet subscription",
                      "type": "HISTORICAL"
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                },
                "examples": {
                  "single": {
                    "summary": "Product DTO",
                    "value": {
                      "id": 1,
                      "name": "men shoes",
                      "categoryId": 5,
                      "unit": "1 pair",
                      "description": "1 pair of leather men shoes",
                      "type": "HISTORICAL"
                    }
                  },
                  "multiple": {
                    "summary": "Count",
                    "value": {
                      "count": 10
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Create the new product.",
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "ProductsController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved products.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                },
                "examples": {
                  "application/json": {
                    "summary": "Product DTO array",
                    "value": [
                      {
                        "id": 1,
                        "name": "man shoes",
                        "categoryId": 5,
                        "unit": "1 pair",
                        "description": "1 pair of leather man shoes",
                        "type": "HISTORICAL"
                      },
                      {
                        "id": 2,
                        "name": "internet",
                        "categoryId": 6,
                        "unit": "monthly subscription",
                        "description": "1 month internet subscription",
                        "type": "HISTORICAL"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "summary": "Return all products.",
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The data to create new product",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              },
              "examples": {
                "single": {
                  "value": {
                    "id": 1,
                    "name": "men shoes",
                    "categoryId": 5,
                    "unit": "1 pair",
                    "description": "1 pair of leather men shoes",
                    "type": "HISTORICAL"
                  }
                },
                "multiple": {
                  "value": [
                    {
                      "id": 1,
                      "name": "men shoes",
                      "categoryId": 5,
                      "unit": "1 pair",
                      "description": "1 pair of leather men shoes",
                      "type": "HISTORICAL"
                    },
                    {
                      "id": 2,
                      "name": "internet",
                      "categoryId": 6,
                      "unit": "monthly subscription",
                      "description": "1 month internet subscription",
                      "type": "HISTORICAL"
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated products.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                },
                "examples": {
                  "single": {
                    "summary": "Single updated product",
                    "value": {
                      "id": 1,
                      "name": "men shoes",
                      "categoryId": 5,
                      "unit": "1 pair",
                      "description": "1 pair of leather men shoes",
                      "type": "HISTORICAL"
                    }
                  },
                  "multiple": {
                    "summary": "Single updated product",
                    "value": [
                      {
                        "id": 1,
                        "name": "men shoes",
                        "categoryId": 5,
                        "unit": "1 pair",
                        "description": "1 pair of leather men shoes",
                        "type": "HISTORICAL"
                      },
                      {
                        "id": 2,
                        "name": "internet",
                        "categoryId": 6,
                        "unit": "monthly subscription",
                        "description": "1 month internet subscription",
                        "type": "HISTORICAL"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "summary": "Update products",
        "tags": [
          "products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved product by id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                },
                "examples": {
                  "application/json": {
                    "summary": "Product DTO",
                    "value": {
                      "id": 1,
                      "name": "men shoes",
                      "categoryId": 5,
                      "unit": "1 pair",
                      "description": "1 pair of leather men shoes",
                      "type": "HISTORICAL"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Return product by id.",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted product by id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                },
                "examples": {
                  "application/json": {
                    "summary": "Product DTO",
                    "value": {
                      "id": 1,
                      "name": "men shoes",
                      "categoryId": 1,
                      "unit": "1 pair",
                      "description": "1 pair of leather men shoes",
                      "type": "HISTORICAL"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Delete product by id.",
        "tags": [
          "products"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The unique identifier of the product"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "categoryId": {
            "type": "number",
            "description": "Category id of the product"
          },
          "unit": {
            "type": "string",
            "description": "Product unit"
          },
          "description": {
            "type": "string",
            "description": "Product description, optional"
          },
          "type": {
            "type": "string",
            "description": "Product Type",
            "enum": [
              "ALL",
              "CURRENT",
              "HISTORICAL"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "categoryId",
          "unit",
          "type"
        ]
      },
...
    }
  }
}